Refactoring: ReadSet in its own module; shared utility functions for tests
